BEGIN HEADERS AND FOOTERS
working on the template files in web > projects > email-blueprints.
Some notes on what Thea did:

I created a copy of modular_templates_pattern.html called
modular_templates_pattern_THEA1.html

In the copy file, I added the same footer from our current monthly
newsletter template, and edited it to fit the width of this template.

I also noted in the file that the header and footer area needs help
with responsive functionality. Apart from finding where the
problematic areas are, I couldn't quite get it to act the same as the
modules. I can show you when we're in the same room.

I didn't get to it but the header logo section should be cleaned up
better since we're moving the social icons in the footer.

So the needs for the template so far are: social icons section
(we'll just have to hard code this part, after yesterday's research),
clean up header and footer, make sure all the responsive stuff is
working, etc.